tilting finally works
I guess this makes sense? It has to account for where the camera is facing to stop it from moving left/right when tilting-only, so getting the camera's .right accounts for that. Vector3.up is just so that you still get the input.x unchanged as the vector's .y as before.

still currently has the bug where tilting makes camera go up/down despite clamped rotation